% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plan.R
\name{analysis_plan.data.frame}
\alias{analysis_plan.data.frame}
\title{Build analysis plan from data}
\usage{
\method{analysis_plan}{data.frame}(.data, .name = "original")
}
\description{
Build analysis plan from data
}
