% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plan.R
\name{analysis_plan}
\alias{analysis_plan}
\title{analysis_plan: This function}
\usage{
analysis_plan(.data, .name)
}
\arguments{
\item{.data}{analysis data}

\item{.name}{starting name for the data}
}
\description{
A description of analysis_plan
}
\details{
Details of analysis_plan
}
