% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep_filter}
\alias{prep_filter}
\title{Filter analysis plan using ids or data}
\usage{
prep_filter(x, ...)
}
\description{
Filter analysis plan using ids or data
}
\examples{
p <- analysis_plan(mtcars)
s <- prep_match(p, a = ~., b = ~dplyr::sample_n(., 2), c = ~dplyr::sample_n(., 2))
# Return analysis with just a
prep_filter(s, nrow(data) > 2)
# Return analysis with a and c
prep_filter(s, nrow(data) > 2 | str_detect(id, 'c'))
# More complex
prep_filter(s, sum(data$vs) > 1)

}
