% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plan.R
\name{analysis_plan}
\alias{analysis_plan}
\title{analysis_plan: This function}
\usage{
analysis_plan(.data, ...)
}
\arguments{
\item{.data}{analysis data}

\item{...}{other arguments}
}
\description{
A description of analysis_plan
}
\details{
A details of analysis_plan
}
